server2/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
pydantic>=1.10.7
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.2
python-dotenv>=0.19.0
joblib>=1.1.0
scapy>=2.4.5
pyyaml>=6.0
apscheduler>=3.10.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
paho-mqtt>=2.2.1
